+++ crates/config/src/project_config.rs
@@ -128,18 +128,12 @@ config_struct!(
         pub language: LanguageType,
 
         /// The layer within the project stack, for categorizing.
-        #[serde(alias = "type")]
         pub layer: LayerType,
 
         /// Defines ownership of source code within the current project, by mapping
         /// file paths and globs to owners. An owner is either a user, team, or group.
         #[setting(nested)]
         pub owners: OwnersConfig,
-
-        /// The default platform for all tasks within the project,
-        /// if their platform is unknown.
-        #[deprecated]
-        pub platform: Option<PlatformType>,
 
         /// Expanded information about the project.
         #[setting(nested)]