+++ packages/types/src/common.ts
@@ -1,5 +1,3 @@
-import type { PlatformType } from './tasks-config';
-
 export type Id = string;
 
 export type Nullable<T> = { [K in keyof T]: T[K] | null };
@@ -7,12 +5,6 @@ export type Nullable<T> = { [K in keyof T]: T[K] | null };
 export interface Duration {
 	secs: number;
 	nanos: number;
-}
-
-export interface Runtime {
-	platform: PlatformType;
-	requirement?: string;
-	overridden?: boolean;
 }
 
 export interface ToolchainSpec {